requirements/common.txt
amqp==1.4.8
anyjson==0.3.3
billiard==3.3.0.22
blinker==1.4
boto==2.39.0
celery==3.1.19
CouchDB==1.0
cv2==1.0
Flask==0.10.1
Flask-Mail==0.9.1
funcsigs==0.4
itsdangerous==0.24
Jinja2==2.8
kombu==3.0.32
MarkupSafe==0.23
mock==1.3.0
numpy==1.10.4
pbr==1.8.1
pep8==1.7.0
picamera==1.10
Pillow==3.1.0
py==1.4.31
pylepton==0.1.2
pytest==2.8.5
pytz==2015.7
RPi.GPIO==0.5.11
six==1.10.0
Werkzeug==0.11.3
wheel==0.24.0
